apple/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
ultralytics==8.3.250
pillow==10.4.0
numpy==2.1.1
